+++ routes/route.js
@@ -1,0 +1,25 @@
+// const inquirer = require('inquirer');
+
+// inquirer
+//   .prompt([
+//     {
+//       type: 'input',
+//       message: 'What is your user name?',
+//       name: 'username',
+//     },
+//     {
+//       type: 'password',
+//       message: 'What is your password?',
+//       name: 'password',
+//     },
+//     {
+//       type: 'password',
+//       message: 'Re-enter password to confirm:',
+//       name: 'confirm',
+//     },
+//   ])
+//   .then((response) =>
+//     response.confirm === response.password
+//       ? console.log('Success!')
+//       : console.log('You forgot your password already?!')
+//   );
