--- schema.sql
@@ -1,17 +1,0 @@
-/* Creating DataBase */
-
-DROP DATABASE IF EXISTS employees_db;
-
-CREATE DATABASE employees_db;
-USE employees_db;
-
-CREATE TABLE department (
-id			    int					PRIMARY KEY NOT NULL AUTO_INCREMENT,
-dept_name	    VARCHAR(30)			NOT NULL
-);
-
-
-/* Seed for the DB */
-
-INSERT INTO department(dept_name)
-VALUES ("sales"),("admin"), ("management");
